+++ src/components/Home/CareerPortal.js
@@ -1,41 +1,8 @@
-/* eslint-disable no-dupe-keys */
 import React from "react";
-import { Grid, Box, Container } from "@material-ui/core";
 
 import { ROUTES } from "src/constants/routes";
-import { makeStyles } from "@material-ui/core/styles";
-import Button from "../../shared/Button";
 import useHomepageData from "../../hooks/queries/homepage";
-import TextSection from "../../shared/TextSection";
-
-const useStyles = makeStyles((theme) => ({
-  root: {
-    position: "relative",
-    height: "460px",
-    borderBottom: "3px solid",
-    borderBottomColor: theme.palette.primary.main,
-    backgroundColor: theme.palette.secondary.dark,
-  },
-
-  container: {
-    position: "absolute",
-    left: "0",
-    top: "64px",
-    [theme.breakpoints.up("xl")]: {
-      left: "calc(12.6vw - 87px)",
-    },
-  },
-
-  image: {
-    position: "absolute",
-    bottom: "0px",
-    right: "24px",
-    [theme.breakpoints.up("xl")]: {
-      right: "calc(12.6vw - 87px)",
-    },
-    height: "70%",
-  },
-}));
+import Separator from "../../shared/Separator";
 
 const CareerPortal = () => {
   const {
@@ -43,33 +10,18 @@ const CareerPortal = () => {
       title,
       body: { body },
       bodyTwo: { bodyTwo },
-      image: {
-        description,
-        file: { url },
-      },
+      image,
     },
   } = useHomepageData();
-  const classes = useStyles();
 
-  return (
-    <Box className={classes.root}>
-      <img src={url} alt={description} className={classes.image} />
-      <Box className={classes.container}>
-        <Container className={classes.textContainer}>
-          <Grid item xs={8}>
-            <TextSection title={title} body={body} />
-          </Grid>
-          <Grid item xs={8}>
-            <Button
-              text={bodyTwo}
-              style={{ marginTop: "24px" }}
-              route={ROUTES.CAREERS}
-            />
-          </Grid>
-        </Container>
-      </Box>
-    </Box>
-  );
+  const data = {
+    route: ROUTES.CAREERS,
+    title,
+    body,
+    bodyTwo,
+    image,
+  };
+  return <Separator {...data} />;
 };
 
 export default CareerPortal;
